+++ example/lib/slides/big_fact_slide.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_deck/flutter_deck.dart';
+
+class BigFactSlide extends FlutterDeckSlideWidget {
+  const BigFactSlide()
+      : super(
+          configuration: const FlutterDeckSlideConfiguration(
+            route: '/big-fact',
+            footer: FlutterDeckFooterConfiguration(showFooter: false),
+          ),
+        );
+
+  @override
+  FlutterDeckSlide build(BuildContext context) {
+    return FlutterDeckSlide.bigFact(
+      title: '100%',
+      subtitle:
+          'The test coverage value that flutter_deck will probably never achieve',
+      theme: FlutterDeckTheme.of(context).copyWith(
+        bigFactSlideTheme: const FlutterDeckBigFactSlideThemeData(
+          titleTextStyle: TextStyle(color: Colors.amber),
+        ),
+      ),
+    );
+  }
+}