+++ example/lib/slides/slides.dart
@@ -1,4 +1,5 @@
 export 'background_slide.dart';
+export 'big_fact_slide.dart';
 export 'blank_slide.dart';
 export 'code_highlight_slide.dart';
 export 'drawer_slide.dart';