+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- feat: add big fact template `FlutterDeckSlide.bigFact`
+
 # 0.5.0
 
 - feat: add `FlutterDeckImageSlideTheme`